requirements.txt
requests == 2.31.0
tqdm == 4.66.1
Pillow == 10.1.0
